Implement comprehensive timestamp synchronization solution
This commit implements a complete solution for timestamp synchronization issues
that were causing video freezing and massive A-V sync differences in RTSP streams.

## Key Changes:

### 1. Unified Timestamp Management
- Add TimestampManager class using IMP hardware timestamps directly
- Ensure all components use the same monotonic timeline
- Eliminate timestamp domain mismatches between audio and video

### 2. Live555 Integration
- Add live555 patch for shared RTP timestamp base support
- Modify build system to automatically apply patch and rebuild live555
- Set shared timestamp base via environment variable in RTSP.cpp
- Eliminate random timestamp base differences between A-V streams

### 3. Enhanced Audio/Video Workers
- Update AudioWorker to use TimestampManager for unified timeline
- Update VideoWorker to use consistent timestamp processing
- Add comprehensive debug logging for timestamp tracking
- Ensure proper timestamp conversion and validation

### 4. Build System Improvements
- Smart patch application logic for live555 dependency
- Automatic live555 clean/rebuild during prudynt builds
- Handle both fresh clones and updates intelligently

## Problem Solved:
- Eliminates massive A-V sync differences (35,791+ seconds)
- Fixes video player freezing and desynchronization issues
- Ensures consistent RTP timestamp bases across all streams
- Provides unified monotonic timeline immune to NTP adjustments

## Technical Details:
- Uses CLOCK_MONOTONIC via IMP_System_GetTimeStamp() for hardware timestamps
- Implements 64-bit nanosecond precision timestamp management
- Applies live555 patch automatically during dependency builds
- Maintains backward compatibility with existing configurations

Fixes video freezing and timestamp synchronization issues in RTSP streaming.

+++ src/AudioWorker.hpp
@@ -36,7 +36,7 @@ private:
     int warnBufferSamplesPerChannel = 0;    // e.g., 3 * targetSamplesPerChannel
 
     // Diagnostics / metrics
-    std::atomic<uint64_t> bufferDropCount{0};
+    std::atomic<uint32_t> bufferDropCount{0};
 };
 
 #endif // AUDIO_SUPPORT
